Edits to materials.tex, cleanup .bib file, create reference.bib.bak
Your references are still a shambles. You need to clean them up.
Journal titles are not all complete, some are abbreviations, others are
not. If you need abbreviations, I suggest you use them. If you need the
full version, use it, but not a mixture!

Also, author names are a mess. Make sure that names or capitalisation
is clean throughout.

Formatting of quotes in references needs cleaning.

Scientific names in your references need cleaning too.

Also, what about a short conclusion? You just end with two figures and
that’s it after M&M. It seems abrupt. For me at least, I’d want a page
summarising what was discussed and what you’ll be doing and it’s
significance.